Fix `compute_projection` for kNN based kernels (#665)
* Fix `compute_projection` for kNN based kernels

The embedding is fixed to use all non-trivial connectivities. As this
requires a connectivity matrix, an `AttributeError` is thrown if the
attribute `_conn` is not set. Previously, only pairs of observations
with non-trivial transition probabilitities were considered.

* Add unit tests for embedding stream plot

Adds the test class `TestProjectionEmbeddingStream` alongside the
corresponding ground truth figures. The test class covers the
`ConnectivityKernel`, `PseudotimeKernel` and `VelocityKernel`.